TRIB-84: Implement notifications api:
notifications.api.service.ts:

Created a mock api that simulates data. It takes two arrays, one for item and the other for message. It builds the notifications with the correct icon:message pairing.

notifications.page.html:

Added an ion List that uses a for loop to create list items. Each item has an onlick event listern to remove the notification from the notification array and display. If no notifications are displayed, it will display "no new notificaion"

notifications.page.ts:

Added logic to handleDelete of the notificaion.

Added a function to get notification icon and display it first in each list.

Added oninit lifecycle to call the api service and get notitications.